requirements.txt
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
torch==2.1.1
torchaudio==2.1.1
transformers==4.35.2
opensmile==2.5.0
pydub==0.25.1
pandas==2.1.4
opencv-python==4.8.1.78
moviepy==1.0.3
openai-whisper==20231117
speechrecognition==3.10.0
ffmpeg-python==0.2.0
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
sentencepiece==0.1.99
protobuf==4.25.1
pyannote.audio==3.1.1
soundfile==0.12.1
